Maven support.
Add a POM for installing the simulation server API
(org.simantics.simulation.scheduling) into a maven repository.  You need
to install it into your local repository with the mvn command: see
the comment in scheduling-pom.xml.  I don't know how to automate it more
than this.  After installation any Maven project should be able to depend
on it in the usual manner.