Convert structure description XML to SCL.
Input to the style sheet is of the form that SearchStructuralFormulas.scl produces.
However, intermediate processing steps are required, e.g., for resolving the variable
case issue (Apros user component properties violate the SCL rule that variables must not
begin with a capital letter). These would appear beyond the capabilities of XSLT 1.0, so we'll
do them in Java or Python.