Parent POM.
This allows Maven configuration sharing.  You'll need to import the
parent pom.xml into Eclipse as a Maven project.  Also import the modules,
except you probably already have them.  Most of the time you'll be
working with the module projects and can largely ignore the parent.
However, any bit of Maven configuration that is common to all
modules should be done via the parent.  Not much is done that way yet.