Inject the time series reader and writer.
OptimisationProblemIO ended up having a lot of static state
because I wanted to cache the mappers (expensive to create),
readers and writers (inexpensive, but while I was at it).  Ugly,
but it seems if you are not using Guice, you'll end up
reinventing it poorly.

CsvTimeSeriesData does not support automatic injection for now.
That would look like another case for assisted inject.

refs #10353: Export functionality