AprosManager log redirection.
Add the log stream as a constructor argument for AprosManager.
In web-client, use log4j-iostreams to redirect the log to Log4j.

I am not entirely happy with the sim-eval API that I ended up with.  The main problem is that the simulation server library only has methods for addding new log streams: there is no easy way to stop logging.  This makes it hard to provide a nicer API in sim-eval.

refs #10466: Redirect AprosManager log